ENCODER FINISHED
Added an option (input to the constructor of the Encoder object) that
allows to use bitwise representation for actions. Also added a function
to print a solution. Also, added some conditions to prevent creating
clauses that will always imply a negated action by having a conflict in
effects or preconds.